+++ src/main/kotlin/com/atlassian/performance/tools/report/api/judge/SampleSizeJudge.kt
@@ -6,19 +6,25 @@ import com.atlassian.performance.tools.report.api.SampleSizeCriteria
 import com.atlassian.performance.tools.report.api.junit.FailedAssertionJUnitReport
 import com.atlassian.performance.tools.report.api.junit.JUnitReport
 import com.atlassian.performance.tools.report.api.junit.SuccessfulJUnitReport
+import com.atlassian.performance.tools.report.api.result.Stats
+import com.atlassian.performance.tools.report.result.PerformanceStats
 
 class SampleSizeJudge {
 
+    @Deprecated("Use the other judge method.")
     fun judge(
         stats: InteractionStats,
         criteria: Map<ActionType<*>, SampleSizeCriteria>
+    ): Verdict = this.judge(
+        stats = PerformanceStats.adapt(stats),
+        criteria = criteria
+    )
+
+    fun judge(
+        stats: Stats,
+        criteria: Map<ActionType<*>, SampleSizeCriteria>
     ): Verdict {
-        val sampleSizes = stats.sampleSizes ?: return Verdict(listOf(
-            FailedAssertionJUnitReport(
-                testName = "Sample size of ${stats.cohort}",
-                assertion = "Sample size results are missing"
-            )
-        ))
+        val sampleSizes = stats.sampleSizes
         val testReports = mutableListOf<JUnitReport>()
         for ((action, sampleCriteria) in criteria) {
             val sampleSize = sampleSizes[action.label] ?: 0
@@ -26,16 +32,16 @@ class SampleSizeJudge {
             if (sampleSize < minimumSampleSize) {
                 val message = "The sample size of ${stats.cohort} ${action.label} is $sampleSize," +
                     " which is under the minimum threshold of $minimumSampleSize"
-                testReports.add(FailedAssertionJUnitReport(testMethodName(stats, action), message))
+                testReports.add(FailedAssertionJUnitReport(testMethodName(stats.cohort, action), message))
             } else {
-                testReports.add(SuccessfulJUnitReport(testMethodName(stats, action)))
+                testReports.add(SuccessfulJUnitReport(testMethodName(stats.cohort, action)))
             }
         }
         return Verdict(testReports)
     }
 
     private fun testMethodName(
-        stats: InteractionStats,
+        cohort: String,
         action: ActionType<*>
-    ) = "Sample_size_for_${stats.cohort}_${action.label}"
+    ) = "Sample_size_for_${cohort}_${action.label}"
 }